feat: VideoCard
