+++ sites/tutors-course-reader/src/pages/Video.svelte
@@ -2,7 +2,7 @@
   import { afterUpdate, getContext } from "svelte";
   import type { Lo } from "tutors-reader-lib/src/types/lo-types";
   import TopicNavigatorCard from "../components/cards/TopicNavigatorCard.svelte";
-  import VideoCard from "../components/cards/VideoCard.svelte";
+  import { VideoCard } from "tutors-ui";
   import type { AnalyticsService } from "../reader-lib/services/analytics-service";
   import { revealSidebar } from "../stores";
   import { talkTransition, viewDelay } from "../components/animations";
